ai_training/requirements.txt
# AI Training Pipeline Requirements
ultralytics>=8.0.200
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=9.0.0
pyyaml>=6.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.11.0
tqdm>=4.64.0
pymupdf>=1.24.2
